pyproject.toml
[project]
name = "ddi-dataset-documentation"
version = "0.1.0"
description = "Dataset documentation builder for DDI display-field sidecars."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "jinja2>=3.1",
  "mkdocs>=1.5",
  "mkdocs-material>=9.5",
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "typer>=0.12",
]

[project.scripts]
ddi_dataset_documentation = "dataset_docs.cli:app"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"
